Refactoring sync.
