+++ integrations/sparse/test_search_pretokenized.py
@@ -17,8 +17,8 @@
 import os
 import shutil
 import unittest
-
 from random import randint
+
 from integrations.lucenesearcher_score_checker import LuceneSearcherScoreChecker
